feat: 모달 컴포넌트 분리, 유효성 검사
