+++ front-end/src/pages/MyPage.jsx
@@ -3,6 +3,7 @@ import axios from "axios";
 import styled from "styled-components";
 
 import blankProfileImage from "../img/blank-profile.png";
+import MyPageModal from "../components/MyPageModal";
 import MyPageTab from "../components/MyPageTab";
 
 const EntireContainer = styled.div`
@@ -82,97 +83,12 @@ const ButtonContainer = styled.div`
   }
 `;
 
-const ModalBackground = styled.div`
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  z-index: 1;
-  backdrop-filter: blur(5px);
-`;
-
-const ModalContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 25rem;
-  height: 35rem;
-  z-index: 999;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  border-radius: 20px;
-  padding: 2rem;
-  gap: 0.6rem;
-  font-size: 1rem;
-  background-color: #fff;
-  > input {
-    height: 2rem;
-    margin-bottom: 0.6rem;
-    border-style: none;
-    border-radius: 5px;
-    padding: 0.5rem;
-    background-color: var(--bg-color);
-  }
-  > button {
-    all: unset;
-    text-align: center;
-    height: 2.5rem;
-    margin-top: 1rem;
-    border-radius: 10px;
-    font-weight: 600;
-    color: #fff;
-    background-color: var(--primary-color);
-    cursor: pointer;
-  }
-`;
-
-const CancleButton = styled.div`
-  display: flex;
-  flex-direction: row-reverse;
-  margin-bottom: 0.4rem;
-  > button {
-    all: unset;
-    font-size: 1.5rem;
-    transform: scaleX(1.5);
-    cursor: pointer;
-    &:hover {
-      color: var(--primary-color);
-    }
-  }
-`;
-
-const ImagebuttonContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  margin: 1rem 0;
-  gap: 1rem;
-  > button {
-    all: unset;
-    text-align: center;
-    height: 2rem;
-    width: 10rem;
-    border-radius: 10px;
-    background-color: var(--sub-color);
-    cursor: pointer;
-  }
-`;
-
 export default function MyPage() {
-  // const [profile, setProfile] = useState({});
-
-  const [member, setMember] = useState({
-    nickName: "",
-    password: "",
-    phone: "",
-    images: "",
-  });
-
   const [modal, setModal] = useState(false);
 
-  // 아마 const memberId = user.memberId 이런 식으로 로그인 시 저장해 둔 유저정보 받아올 듯
+  // const [profile, setProfile] = useState({});
+
+  // get 요청 부분
   // useEffect(() => {
   //   const fetchMember = async () => {
   //     const response = await axios.get(`${process.env.REACT_APP_BASE_URL}/members/${memberId}`);
@@ -180,30 +96,60 @@ export default function MyPage() {
   //   };
   //   fetchMember();
   // }, []);
+  // 아마 const memberId = user.memberId 이런 식으로 로그인 시 저장해 둔 유저정보 받아올 듯
+  const memberId = 1;
   const profile = {
     memberId: 1,
     email: "test@gmail.com",
     nickName: "nickname",
     phone: "010-1234-5678",
-    image: "",
+    images: "https://cdn.pixabay.com/photo/2016/03/04/22/54/animal-1236875_1280.jpg",
+    memberStatus: "ACTIVE",
+    createdDate: "2023-05-12T14:28:14.621685",
+    updateDate: "2023-05-12T14:29:16.909628",
   };
 
-  const { email, nickName, phone, image } = profile;
+  const { email, nickName, phone, images } = profile;
+
+  const [member, setMember] = useState({
+    nickName,
+    password: "",
+    phone,
+    images,
+  });
+  console.log(member);
+
+  const onChange = event => {
+    const { name, value } = event.target;
+    setMember(previous => ({ ...previous, [name]: value }));
+  };
+
+  // patch 요청 부분
+  const onSubmit = async event => {
+    event.preventDefault();
+    try {
+      await axios.patch(`${process.env.REACT_APP_BASE_URL}/members/${memberId}`, member);
+      alert("회원 정보가 수정되었습니다.");
+    } catch (error) {
+      alert("회원 정보 수정에 실패했습니다.");
+      console.log(error);
+    }
+  };
 
   return (
     <EntireContainer>
       <ProfileSection>
-        {image === "" || image === null ? (
+        {images === "" || images === null ? (
           <img src={blankProfileImage} alt="blanked user profile" />
         ) : (
-          <img src={image} alt="user profile" />
+          <img src={images} alt="user profile" />
         )}
         <ProfileInformation>
           <span className="nickName">{nickName}</span>
           <span>{email}</span>
           <span>{phone}</span>
           <ButtonContainer>
-            {modal && <Modal setModal={setModal} nickName={nickName} phone={phone} />}
+            {modal && <MyPageModal setModal={setModal} member={member} onSubmit={onSubmit} onChange={onChange} />}
             <button type="button" onClick={() => setModal(true)}>
               수정하기
             </button>
@@ -215,62 +161,5 @@ export default function MyPage() {
       </ProfileSection>
       <MyPageTab />
     </EntireContainer>
-  );
-}
-
-function Modal({ setModal, nickName, phone }) {
-  const labels = [
-    {
-      id: "nickName",
-      title: "닉네임",
-      children: <input type="text" id="nickName" placeholder={nickName} minLength="2" maxLength="8" required />,
-    },
-    {
-      id: "phone",
-      title: "휴대폰 번호",
-      children: <input type="tel" id="phone" placeholder={phone} required />,
-    },
-    {
-      id: "password",
-      title: "비밀번호",
-      children: (
-        <input
-          type="password"
-          id="password"
-          placeholder="비밀번호를 입력하세요."
-          pattern="/^(?=.*[a-zA-Z])(?=.*[!@#$%^*+=-])(?=.*[0-9])(?=\S+$).{8,16}$/"
-          title="영문,숫자,특문을 조합해서 8자 이상 입력해주세요"
-          required
-        />
-      ),
-    },
-    {
-      id: "passwordCheck",
-      title: "비밀번호 확인",
-      children: <input type="password" id="passwordCheck" placeholder="비밀번호를 다시 한 번 입력해주세요." required />,
-    },
-  ];
-
-  return (
-    <ModalBackground>
-      <ModalContainer>
-        <CancleButton>
-          <button type="button" onClick={() => setModal(false)}>
-            X
-          </button>
-        </CancleButton>
-        <ImagebuttonContainer>
-          <button type="button">이미지 등록하기</button>
-          <button type="button">이미지 삭제하기</button>
-        </ImagebuttonContainer>
-        {labels.map(label => (
-          <>
-            <label htmlFor={label.id}>{label.title}</label>
-            {label.children}
-          </>
-        ))}
-        <button type="button">수정하기</button>
-      </ModalContainer>
-    </ModalBackground>
   );
 }
